+++ metacells/pipeline/direct.py
@@ -78,6 +78,8 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
     feature_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
     forbidden_gene_names: Optional[Collection[str]] = None,
     forbidden_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
+    bystander_gene_names: Optional[Collection[str]] = None,
+    bystander_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
     feature_correction: Optional[FeatureCorrection] = None,
     cells_similarity_value_normalization: float = pr.cells_similarity_value_normalization,
     cells_similarity_log_data: bool = pr.cells_similarity_log_data,
@@ -164,6 +166,10 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
             A boolean mask of genes which are forbidden from being chosen as "feature" genes based
             on their name.
 
+        ``bystander_gene``
+            A boolean mask of genes which are not only forbidden from being chosen as "feature" genes, but also are
+            ignored when computing deviant (outlier) cells, based on their name.
+
         ``feature_gene``
             A boolean mask of the "feature" genes.
 
@@ -205,10 +211,10 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
        ``feature_min_gene_total`` (default: {feature_min_gene_total}), ``feature_min_gene_top3``
        (default: {feature_min_gene_top3}), ``feature_min_gene_relative_variance`` (default:
        {feature_min_gene_relative_variance}), ``feature_gene_names`` (default:
-       {feature_gene_names}), ``feature_gene_patterns`` (default: {feature_gene_patterns}),
-       ``forbidden_gene_names`` (default: {forbidden_gene_names}), ``forbidden_gene_patterns``
-       (default: {forbidden_gene_patterns}) and ``random_seed`` (default: {random_seed}) to make
-       this replicable.
+       {feature_gene_names}), ``feature_gene_patterns`` (default: {feature_gene_patterns}), ``forbidden_gene_names``
+       (default: {forbidden_gene_names}), ``forbidden_gene_patterns`` (default: {forbidden_gene_patterns})
+       ``bystander_gene_names`` (default: {bystander_gene_names}), ``bystander_gene_patterns`` (default:
+       {bystander_gene_patterns}) and ``random_seed`` (default: {random_seed}) to make this replicable.
 
     2. Apply the ``feature_correction`` function, if any, to modify the downsampled features data.
 
@@ -289,8 +295,8 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
         min_gene_top3=feature_min_gene_top3,
         forced_gene_names=feature_gene_names,
         forced_gene_patterns=feature_gene_patterns,
-        forbidden_gene_names=forbidden_gene_names,
-        forbidden_gene_patterns=forbidden_gene_patterns,
+        forbidden_gene_names=list(forbidden_gene_names or []) + list(bystander_gene_names or []),
+        forbidden_gene_patterns=list(forbidden_gene_patterns or []) + list(bystander_gene_patterns or []),
         random_seed=random_seed,
     )
 
@@ -391,12 +397,10 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
         deviant_votes_of_cells = np.zeros(adata.n_obs, dtype="float32")
         dissolved_of_cells = np.zeros(adata.n_obs, dtype="bool")
 
+        tl.find_named_genes(adata, names=bystander_gene_names, patterns=bystander_gene_patterns, to="bystander_gene")
         ut.set_v_data(adata, "gene_deviant_votes", deviant_votes_of_genes, formatter=ut.mask_description)
-
         ut.set_o_data(adata, "cell_deviant_votes", deviant_votes_of_cells, formatter=ut.mask_description)
-
         ut.set_o_data(adata, "dissolved", dissolved_of_cells, formatter=ut.mask_description)
-
         ut.set_o_data(adata, "metacell", candidate_of_cells, formatter=ut.groups_description)
 
     else:
@@ -407,6 +411,8 @@ def compute_direct_metacells(  # pylint: disable=too-many-statements,too-many-br
             abs_folds=deviants_abs_folds,
             max_gene_fraction=deviants_max_gene_fraction,
             max_cell_fraction=deviants_max_cell_fraction,
+            bystander_gene_names=bystander_gene_names,
+            bystander_gene_patterns=bystander_gene_patterns,
         )
 
         tl.dissolve_metacells(